worked on many-to-many relationships, dto-s

+++ src/main/java/com/musicians/d2ovj9haladojavalev/controller/AlbumController.java
@@ -2,7 +2,10 @@ package com.musicians.d2ovj9haladojavalev.controller;
 
 import com.musicians.d2ovj9haladojavalev.dto.AlbumDTO;
 import com.musicians.d2ovj9haladojavalev.entity.Album;
+import com.musicians.d2ovj9haladojavalev.entity.Artist;
+import com.musicians.d2ovj9haladojavalev.exception.DataNotFoundException;
 import com.musicians.d2ovj9haladojavalev.service.AlbumService;
+import com.musicians.d2ovj9haladojavalev.service.ArtistService;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.validation.BindingResult;
@@ -10,15 +13,18 @@ import org.springframework.web.bind.annotation.*;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.NoSuchElementException;
 
 @RestController
 @RequestMapping("/api")
 public class AlbumController {
     private final AlbumService albumService;
+    private final ArtistService artistService;
 
     @Autowired
-    public AlbumController(AlbumService albumService) {
+    public AlbumController(AlbumService albumService, ArtistService artistService) {
         this.albumService = albumService;
+        this.artistService = artistService;
     }
 
     @GetMapping("/album")
@@ -42,12 +48,29 @@ public class AlbumController {
         if (bindingResult.hasErrors()) {
             System.out.println(bindingResult.getAllErrors());
         }
+        if (album.getArtist() != null) {
+            try {
+                Artist artist = artistService.getArtist(album.getArtist().getId());
+                album.setArtist(artist);
+            } catch (NoSuchElementException e){
+                throw new DataNotFoundException(
+                        "Whoops. Cannot find artist with id:" +
+                                album.getArtist().getId()
+                );
+            }
+        }
         albumService.addAlbum(album);
         return album;
     }
 
     @DeleteMapping("/album/{id}")
-    public void deleteAlbumById(@PathVariable Long id) {
-        albumService.deleteAlbumById(id);
+    public String deleteAlbumById(@PathVariable Long id) {
+        Album album = albumService.getAlbumById(id);
+        if (album == null) {
+            throw new DataNotFoundException("Whoops! Album with id " + id + " not found.");
+        } else {
+            albumService.deleteAlbumById(id);
+        }
+        return "Artist deleted with id: " + id;
     }
 }
